Moved cloud computing transfer data functionality to cloud_API.py
